+++ scopes/scope/scope/ui/components-drawer/components-drawer.module.scss
