+++ scopes/scope/scope/ui/components-drawer/components.drawer.tsx
@@ -10,7 +10,7 @@ import { Text } from '@teambit/base-ui.text.text';
 import { mutedItalic } from '@teambit/ui.styles.muted-italic';
 import { ellipsis } from '@teambit/ui.styles.ellipsis';
 import { useScope } from '../use-scope';
-import styles from './components.drawer.module.scss';
+import styles from './components-drawer.module.scss';
 
 export class ComponentsDrawer implements DrawerType {
   constructor(private treeNodeSlot: ComponentTreeSlot) {}
