+++ scopes/workspace/workspace/ui/workspace-components-drawer/workspace-components-drawer.module.scss
